backend/requirements_pinned.txt
openai==1.35.7
google-generativeai==0.7.2
python-dotenv==1.0.1

pyttsx3==2.90
SpeechRecognition==3.10.0
pyaudio==0.2.13

colorama==0.4.6
requests==2.31.0
flask==3.0.3
flask-cors==4.0.0
pyserial==3.5

transformers==4.40.2
nltk==3.8.1
